New modules were added
