+++ nncf/onnx/graph/node_utils.py
@@ -14,14 +14,11 @@
 import onnx
 import numpy as np
 
-from nncf.common.graph.transformations.commands import TargetType
 from nncf.common.graph.graph import NNCFNode
 from nncf.onnx.graph.metatypes.onnx_metatypes import LAYERS_WITH_BIAS_METATYPES
 from nncf.onnx.graph.onnx_graph import ONNXGraph
 from nncf.onnx.graph.metatypes.onnx_metatypes import ONNX_OPERATION_METATYPES
 from nncf.onnx.graph.metatypes.onnx_metatypes import ONNXIdentityMetatype
-from nncf.onnx.graph.transformations.commands import ONNXTargetPoint
-from nncf.onnx.graph.transformations.commands import ONNXBiasCorrectionCommand
 
 
 def is_node_with_bias(node: NNCFNode) -> bool:
@@ -50,14 +47,3 @@ def get_bias_value(node: NNCFNode, model: onnx.ModelProto) -> np.ndarray:
     if metatype == ONNXIdentityMetatype:
         return onnx_graph.get_initializers_value(node.input[0])
     raise RuntimeError('Could not find the bias value of the node')
-
-
-def create_bias_correction_command(node: NNCFNode, bias_value: np.ndarray) -> ONNXBiasCorrectionCommand:
-    """
-    :param node:
-    :param bias_value:
-    :return:
-    """
-    bias_port_id = node.metatype.weight_definitions.bias_port_id
-    target_point = ONNXTargetPoint(TargetType.LAYER, node.node_name, bias_port_id)
-    return ONNXBiasCorrectionCommand(target_point, bias_value)
